Move post object to common library
