+++ api/src/main/java/com/minibcr/api/ApiApplication.java
@@ -3,7 +3,8 @@ package com.minibcr.api;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-@SpringBootApplication public class ApiApplication {
+@SpringBootApplication(scanBasePackages = {"com.minibcr.*"})
+public class ApiApplication {
 
     public static void main(String[] args) {
         SpringApplication.run(ApiApplication.class, args);
