+++ api/src/main/java/com/minibcr/api/PostController.java
@@ -12,6 +12,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.minibcr.commons.Post;
+
 @RestController()
 @RequestMapping("/post")
 public class PostController {
