+++ api/src/main/java/com/minibcr/api/PostServiceImpl.java
@@ -5,6 +5,8 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.minibcr.commons.Post;
+
 @Service
 public class PostServiceImpl implements PostService {
