+++ commons/src/main/java/com/minibcr/commons/CommonsApplication.java
@@ -1,0 +1,17 @@
+package com.minibcr.commons;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@SpringBootApplication
+public class CommonsApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(CommonsApplication.class, args);
+    }
+
+}
