+++ commons/src/test/java/com/minibcr/commons/CommonsApplicationTests.java
@@ -1,0 +1,13 @@
+package com.minibcr.commons;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CommonsApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
