+++ crawler/src/main/java/com/minibcr/crawler/Crawler.java
@@ -7,6 +7,8 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
+import com.minibcr.commons.Post;
+
 @Component
 @EnableScheduling
 @PropertySource("classpath:crawler.properties")
@@ -14,16 +16,16 @@ public class Crawler {
 
     private final YcombinatorService ycombinatorService = new YcombinatorService();
     private final YcombinatorConfig ycombinatorConfig = new YcombinatorConfig();
-
-
+    
     @Scheduled(fixedRate = 30000)
     public void run() {
 
         String url = ycombinatorConfig.getUrl();
         String[] disallowed = ycombinatorConfig.getDisallowed();
 
+
         try {
-            for (YcombinatorPost post : ycombinatorService.getArticles(url, disallowed)) {
+            for (Post post : ycombinatorService.getArticles(url, disallowed)) {
                 System.out.println("-----------");
                 System.out.println("Title: " + post.getTitle());
                 System.out.println("Link: " + post.getLink());
