+++ crawler/src/main/java/com/minibcr/crawler/YcombinatorService.java
@@ -4,10 +4,9 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.jsoup.Jsoup;
-import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
-import org.jsoup.select.Elements;
+
+import com.minibcr.commons.Post;
 
 public class YcombinatorService {
     public YcombinatorService() {
@@ -16,8 +15,8 @@ public class YcombinatorService {
     private static YcombinatorPostBuilder ycombinatorPostBuilder = new YcombinatorPostBuilder();
     private static YcombinatorClient ycombinatorClient = new YcombinatorClient();
 
-    public List<YcombinatorPost> getArticles(String url, String[] disallowed) throws IOException {
-        List<YcombinatorPost> ycombinatorPosts = new ArrayList<YcombinatorPost>();
+    public List<Post> getArticles(String url, String[] disallowed) throws IOException {
+        List<Post> ycombinatorPosts = new ArrayList<>();
 
         if (disallowed(url, disallowed)) {
             System.out.println("This url is disallowed: " + url);
