+++ crawler/src/main/resources/application.properties
@@ -1,1 +1,8 @@
+spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.PostgreSQLDialect
+spring.jpa.hibernate.ddl-auto=none
+spring.jpa.hibernate.show-sql=true
+spring.datasource.url=jdbc:postgresql://localhost:5432/postgres
 
+spring.sql.init.mode=always
+spring.datasource.initialize=true
+spring.sql.init.continue-on-error=true
